requirements.txt
fastapi
uvicorn
requests
pydantic
python-multipart
PyJWT
python-dotenv